Move form previews so they use the same logical paths as lookbook
