+++ lib/primer/static/generate_previews.rb
@@ -11,7 +11,7 @@ module Primer
       class << self
         def call
           Lookbook.previews.filter_map do |preview|
-            next if preview.preview_class == Primer::Forms::FormsPreview
+            next if preview.preview_class == Primer::FormsPreview
             next if Primer::Accessibility.ignore_preview?(preview.preview_class)
 
             component = preview.components.first&.component_class
